make ShipHub-Local work for my account
I was getting crashes on launch because I have a bunch of disabled repos
associated with my account.  The disabled repos are from an old GH org
that no longer has active billing.

For these disabled repos, we'd get unexpected 404 responses when
fetching the assignees.  The JSON would have message = "Not found" and
none of the fields we expected.

To workaround, we'll filter out repos that return 404 when fetching
assignees.