IssueWeb is automatically built as part of ShipHub
If we don't find the node_modules dir, we run npm install to generate
it.

If we don't find the dist dir OR if git status reports any changed
files or new untracked files in the IssueWeb dir, we re-run webpack to
generate the bundle.

The result is...

 - First ever builds are always slow because we have to install modules
   and run webpack.

 - Incremental builds (IFF you don't change anything in IssueWeb) are
   fast since we correctly skip all webpacking.

 - Incremental builds (if you've changed files in IssueWeb) take about
   5 or 6 seconds longer because we run webpack each time.  But, if
   you're working on IssueWeb, maybe use the new `UseWebpackDevServer`
   option I talked about to get live editing while ShipHub is running.
   (To be added in a later PR).

I thought about doing that Makefile stuff, but also thought it would
take me a really long time.  I'm not a Makefile pro.  And, the potential
input file list is huge (our code, plus every node module we include).
This felt like a good compromise.