for ref + x-ref events, only store fields we need
For referenced and cross-referenced events, we have to fetch referenced
issues, pull requests, or commits so we can show the right UI.

I used to store the entire responses of those extra requests in the JSON
blob for the event.  That causes a lot of bloat - commits and pull
requests include entire patch contents which can be huge.

Now we'll only store the specific fields we need.  All are prefixed with
"ship_" right now so they're easy to find.