Fix realartists/shiphub-cocoa#131 Enable spellcheck (if possible)
Create a CodeMirror add-on to do spellchecking using the system
spellchecker. The add-on hands the CodeMirror contents off to
IssueViewController, who in turn background processes those contents
with NSSpellChecker. When results are found, they are sent back to
javascript to be turned into CodeMirror marks, which add a CSS class
to render the red underlines for the text. Special care is taken to
avoid showing red underlines for text you're currently typing (this
matches the built in NSTextView behavior).

When showing the context menu for a misspelling, we first pass the
misspelling's context (text, and location in document) up into the
native code in IssueViewController where the actual spell checking
suggestions are retrieved and shown in an NSMenu. If one of the
suggestions is chosen, the replacement word and the context is sent
back to javascript to do replacement in the CodeMirror instance.