make syncing work for my account again
My account has an organization that's no longer active, and when you try
to fetch the assignees for that org's repos, you'll get a 404.  This was
preventing GHSyncConnection from ever sync'ing my repos.

When we get an error response on these dead repos, we don't want the
entire jsonTasks job to fail just because one fetch failed.  We should
instead return all results + errors and let the caller prune out the
failures.

cc @james-howard