fix issue where event handlers accumulate on the typeahead element
This updateTypeahead function is called every time componentDidUpdate is
called.  As the component would update, we'd end up subscribing over and
over again to the same events and those would accumulate.  e.g., when
the Enter key was pressed, we might handle it 4-5 times.

I don't know for sure this was causing any problems, but it's at least
making my debugging more noisy.

The react-ish way to solve this would be to move all these subscriptions
to componentDidMount.  But, whatevs - this thing is an odd beast anyway
in that we have to destroy/recreate on every update.  off() will do for
us.