fix bug where "New Label: <input>" would briefly appear in label field
If you typed "blah blah" into the label field and hit <Enter>, you would
briefly see "New Label: blah blah" show in the input box as the label
sheet opened.

This would happen because we had the matcher returning "New Label:
<input>" as one of the possible entries.  When <Enter> was pressed;
completer would call completeOrFail() to set the input value to the true
matched value.  That matched value was the whole string "New Label:
<input>".

To fix, instead of the matcher adding "New Label: <input>" to the
matches, we'll just add "<input>".  This easier to understand, anyway.