Improve sync performance.
A two-level cache for Core Data is used to improve sync
performance.

The first level is an intra syncConnection:receivedEntries:... cache
that stores a mapping between EntityName.identifier and actual
NSManagedObjects, where those objects are either already inserted or
pending insertion.

The second level is a DataStore wide cache from EntityName.identifier
to NSManagedObjectID. If we miss the first cache but hit this second
cache, we still have to go out to the db, but it will be via the
faster -objectWithID rather than via a fetch request.