Improve sync performance for large databases even further.
Instead of finding objects we need by identifier one by one and then
caching them in a two level cache like I did in my patch yesterday,
instead do an up front processing step for each bundle of sync data in
syncConnection:receivedEntries: to identify all of the database
objects we will need. This means finding all of the entity types and
the corresponding sets of identifiers that we will encounter for the
remainder of the sync operation.

Once we have those identifiers known, fetch all of the known objects
with those identifiers from the database in a single call per entity
type, and cache them. Then, let the sync process proceed as usual,
hitting the _syncCache for any objects needed, as opposed to doing
NSFetchRequests for them.

Additionally, do some optimizations to MetadataStore. In particular,
make sorting allAssignees lazy, as this data is rarely accessed.
Additionally, avoid using a fetch request to load organization members
and instead prefer to traverse via CoreData accessors for this
information, as that's much faster. Finally, make updating the
MetadataStore a synchronous operation when metadata changes in _moc,
as this is faster and technically needed for correctness (loading
Issues and other data does depend on having up-to-date metadata).