second pass at sending analytics events from the client
This is the new version with no long-running timers.  There's only a
timer running for as long as the queue is non-empty.

Events are added to a queue.  Whenever an event is added, if not already
scheduled, we'll the schedule the queue to be flushed 60 seconds from
then.  After we finish flushing the queue, if more events have been
added since we began, we schedule another flush 60 seconds later.

The queue is saved to disk when app is quit and restored on app start.

Seems to work.

Here's a sampling of events coming from my local dev --
https://mixpanel.com/report/1156006/live/

This diff doesn't add that many events, but it's at least enough to see
a funnel on the login flow.  We'll add more events in later commits.

Closes #366