fix issue where username is not attached to crash reports
We already implement the delegate method to do this - it's just that we
always end up returning nil for the username because the `_auth` ivar is
not yet initialized.

Crash reports get sent as soon as we call `startManager` on
`BITHockeyManager` and we do that very early in startup.  That's good -
we wouldn't want to delay that, because I think startManager registers
the uncaught exception handler.

Let's pull from NSUserDefaults instead.

Closes #445