Fix PR diff view to show only the diff between base and a hypothetical base post-merge.
Previously, we were showing the full diff between base and head in the
PR, which was incorrect in the following case:

1. Create a branch off master (or whatever branch you want)
2. Add commits to master (or whatever base branch)
3. Add commits to branch created in step 1
4. Create pull request from branch

In this case, the diff from base to branch would include the revert of
the changes made in step 2 above. This isn't useful, and it isn't what
GitHub shows. Instead, what we conceptually want to do is this:

git checkout master
git checkout -b TestMergePR
git merge PR
git diff master...TestMergePR

Of course, this isn't what I actually do, but it's conceptually it.
What I actually do is just find the parent of the first commit in the
PR, and then find the tree diff from that to the head commit in the
PR.