Allow DataStore to execute complex subqueries that Core Data lacks support for.
Suppose we want to execute the following predicate:

NOT ANY labels.name BEGINSWITH[c] "area-"

We can't execute that directly because Core Data will generate incorrect SQL and return the wrong data. So we rewrite it:

count:(subquery(labels.name, $x, $x BEGINSWITH[c] "area-")) == 0

We can't execute that either, because Core Data only knows how to generate SQL for the following operators in that construction: =, !=, and IN (though it does allow any combination of those via compound predicates). It does not know how to generate SQL for BEGINSWITH, CONTAINS, OR MATCHES operators (likely because these require calling back into native code from sqlite via custom functions).

So, we break the query up into two parts. First, we ask for all of the label managed objects that match the subquery predicate, then we rewrite the query to a construction that is both fast and possible for core data to execute:

count:(subquery(labels, $x, $x IN {...})) == 0

Where the {...} is the result of an additional fetch request to fetch all labels matching the predicate:

name BEGINSWITH[c] "area-"

Extrapolate and recurse on these ideas, and you have the gist of this commit. :)